chore: fix clippy

+++ src/decoder.rs
@@ -51,7 +51,7 @@ impl Decoder {
         memory: &mut M,
         pc: u64,
     ) -> Result<u32, Error> {
-        let mut instruction_bits = u32::from(memory.execute_load32(pc)?);
+        let mut instruction_bits = memory.execute_load32(pc)?;
         if instruction_bits & 0x3 != 0x3 {
             instruction_bits &= 0xffff;
         }
